+++ packages/components/src/index.ts
@@ -44,3 +44,4 @@ export { DownloadFiles } from './DownloadFiles/DownloadFiles'
 export { ResearchEditorOverview } from './ResearchEditorOverview/ResearchEditorOverview'
 export type { ResearchEditorOverviewUpdate } from './ResearchEditorOverview/ResearchEditorOverview'
 export { FollowResearchButton } from './FollowResearchButton'
+export { Tooltip } from './Tooltip/Tooltip'
